+++ packages/terra-application/tests/jest/application-navigation/ApplicationNavigation.test.jsx
@@ -1,10 +1,22 @@
 import React from 'react';
-
+import { shallowWithIntl } from '@cerner/terra-enzyme-intl';
 import ApplicationNavigation from '../../../src/application-navigation/ApplicationNavigation';
+
+jest.mock('react-intl', () => {
+  const reactIntl = jest.requireActual('react-intl');
+  const intl = reactIntl.createIntl({
+    locale: 'en',
+  });
+
+  return {
+    ...reactIntl,
+    useIntl: () => intl,
+  };
+});
 
 describe('ApplicationNavigation', () => {
   it('should render default element', () => {
-    const shallowComponent = shallow(
+    const shallowComponent = shallowWithIntl(
       <ApplicationNavigation
         id="test-application-id"
       />,
@@ -14,7 +26,7 @@ describe('ApplicationNavigation', () => {
   });
 
   it('should render with all props', () => {
-    const shallowComponent = shallow(
+    const shallowComponent = shallowWithIntl(
       <ApplicationNavigation
         id="test-application-id"
         activeBreakpoint="large"