+++ packages/terra-application/tests/jest/navigation-prompt/getUnsavedChangesPromptOptions.test.jsx
@@ -1,4 +1,4 @@
-import { mockIntl } from 'terra-enzyme-intl';
+import { mockIntl } from '@cerner/terra-enzyme-intl';
 
 import getUnsavedChangesPromptOptions from '../../../src/navigation-prompt/getUnsavedChangesPromptOptions';