+++ packages/terra-application-docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Upgraded to intl v5
 
 ## 2.7.0 - (February 14, 2022)